NonFollowers(GitHub) - Py/requirements.txt
beautifulsoup4==4.11.2
python-dotenv==1.0.0
requests==2.25.1
selenium==4.8.2
